connect address with orders and change/add column

+++ client/orders.php
@@ -6,7 +6,15 @@ $userid = $_SESSION['userid'];
 $query = "INSERT INTO `orders` (userid, pid,seller_id) SELECT `userid`, `pid`, `seller_id` FROM `cart` WHERE `userid`=$userid";
 $result = mysqli_query($conn, $query);
 
-if ($result) {
+$address_query = "SELECT `address_id` FROM `addresses` WHERE `userid`=$userid";
+$address_result = mysqli_query($conn, $address_query);
+$address_row = mysqli_fetch_assoc($address_result);
+$address_id = $address_row['address_id'];
+// print_r($address_id);
+$update_query = "UPDATE `orders` SET `address_id`=$address_id WHERE `userid`=$userid";
+$update_table = mysqli_query($conn, $update_query);
+
+if ($result && $update_table) {
     $delete_query = "DELETE FROM `cart` WHERE `userid`=$userid";
     $deleted_result = mysqli_query($conn, $delete_query);
     if ($deleted_result) {
